requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-multipart==0.0.20
httpx==0.28.1
beautifulsoup4==4.12.3
anthropic==0.42.0
google-generativeai==0.8.6
python-dotenv==1.2.2
pydantic==2.10.4
jinja2==3.1.5
python-docx==1.1.2
PyPDF2==3.0.1
websockets==14.1
google-auth==2.38.0
PyJWT==2.12.1
